# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_rekall_backend_internal_interfaces_http_dto.ErrorBody:
    properties:
      code:
        example: CALL_NOT_FOUND
        type: string
      details: {}
      message:
        example: call with the given ID was not found
        type: string
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorBody'
      success:
        example: false
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.Meta:
    properties:
      page:
        example: 1
        type: integer
      per_page:
        example: 20
        type: integer
      total:
        example: 150
        type: integer
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponse'
        type: array
      meta:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.Meta'
      success:
        example: true
        type: boolean
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.UserResponse:
    properties:
      created_at:
        example: "2026-01-15T09:00:00Z"
        type: string
      email:
        example: alice@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      full_name:
        example: Alice Nguyen
        type: string
      id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
      role:
        enum:
        - admin
        - member
        example: member
        type: string
    type: object
  github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponse'
      success:
        example: true
        type: boolean
    type: object
...
  internal_interfaces_http_handlers.CreateUserRequest:
    properties:
      email:
        example: bob@example.com
        type: string
      full_name:
        example: Bob Smith
        maxLength: 255
        minLength: 1
        type: string
      role:
        enum:
        - admin
        - member
        example: member
        type: string
    required:
    - email
    - full_name
    type: object
host: localhost:8080
info:
  contact:
    email: engineering@rekall.io
    name: Rekall Engineering
  description: |-
    Rekall is a call intelligence platform. Record, transcribe, and analyse voice and video calls.
    This API powers the Rekall web application and is available for direct integration.

    ## Authentication
    Protected endpoints require a JWT access token obtained from `POST /api/v1/auth/login`.
    Pass the token in the `Authorization` header as `Bearer <token>`.
    The refresh token is transported exclusively via an `HttpOnly` cookie set by the server.

    ## Response Envelope
    All responses are wrapped in a standard envelope:
    - Success: `{ "success": true, "data": <T>, "meta": <pagination|null> }`
    - Error:   `{ "success": false, "error": { "code": "ERROR_CODE", "message": "Human readable message", "details": <null|object> } }`

    ## Error Codes
    Every error response includes a machine-readable `code` field. Common codes:
    `INVALID_CREDENTIALS`, `EMAIL_NOT_VERIFIED`, `EMAIL_ALREADY_REGISTERED`,
    `INVALID_OR_EXPIRED_TOKEN`, `INVALID_REFRESH_TOKEN`, `MISSING_TOKEN`, `INVALID_TOKEN`,
    `CALL_NOT_FOUND`, `USER_NOT_FOUND`, `SLUG_ALREADY_TAKEN`, `ALREADY_A_MEMBER`,
    `CANNOT_REMOVE_OWNER`, `INVITATION_EMAIL_MISMATCH`, `DEPARTMENT_NOT_FOUND`, `FORBIDDEN`
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Rekall API
  version: 1.0.0
paths:
...
  /api/v1/users:
    get:
      description: Returns a paginated list of all platform users. **Requires `admin`
        role.** Regular members receive 403.
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated user list
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserListResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Insufficient role — admin required (FORBIDDEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users  [admin]
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new platform user directly (admin operation — bypasses
        the self-registration flow). **Requires `admin` role.**
      parameters:
      - description: User payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Insufficient role — admin required (FORBIDDEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "409":
          description: Email already registered (EMAIL_ALREADY_REGISTERED)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a user  [admin]
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Soft-deletes a platform user. **Requires `admin` role.**
      parameters:
      - description: User UUID
        example: 00000000-0000-0000-0000-000000000001
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User deleted
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Insufficient role — admin required (FORBIDDEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "404":
          description: User not found (USER_NOT_FOUND)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user  [admin]
      tags:
      - Users
    get:
      description: Returns a single user profile by UUID. **Requires `admin` role.**
      parameters:
      - description: User UUID
        example: 00000000-0000-0000-0000-000000000001
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.UserResponseEnvelope'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Insufficient role — admin required (FORBIDDEN)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "404":
          description: User not found (USER_NOT_FOUND)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user  [admin]
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT access token as: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"